Resolve circular dependency in tendermint-rpc
