+++ packages/iov-tendermint-rpc/types/adaptor.d.ts
@@ -46,10 +46,3 @@ export interface Responses {
   readonly decodeNewBlockHeaderEvent: (response: SubscriptionEvent) => responses.NewBlockHeaderEvent;
   readonly decodeTxEvent: (response: SubscriptionEvent) => responses.TxEvent;
 }
-/**
- * Returns an Adaptor implementation for a given tendermint version.
- * Throws when version is not supported.
- *
- * @param version full Tendermint version string, e.g. "0.20.1"
- */
-export declare function adaptorForVersion(version: string): Adaptor;
